app/requirements.txt
Flask==3.1.2
pandas==2.3.3
psycopg2-binary==2.9.11
SQLAlchemy==2.0.46
marshmallow==4.2.1
pytest==9.0.2
flask-cors==6.0.2
flask-swagger-ui==5.21.0
python-dotenv==1.2.1